changes following CR
